Add files via upload
Distributed protocols for computing consensus. The python program takes as input the number of nodes (n) and the minority fraction. The goal is to run a distributed protocol (in the population protocol model, where each nodes are matched uniformly at random in each step) and to eventually reach consensus with the correct initial majority bit. 

The program implements the repeated majority protocol (see references in the README file) and plots the evolution of the system over time.